Change the action cable implementation
